+++ frontend/package.json
@@ -8,15 +8,15 @@
   "scripts": {
     "all-checks": "npm run test && npm run lint && npm run ts:check && npm run build",
     "build": "next build",
-    "dev": "next dev",
-    "dev:nr": "NODE_OPTIONS='-r @newrelic/next' next dev",
+    "dev": "NEW_RELIC_ENABLED=false next dev",
+    "dev:nr": "NODE_OPTIONS='-r newrelic' next dev",
     "debug": "NODE_OPTIONS='--inspect' next dev",
     "format": "prettier --write '**/*.{js,json,md,ts,tsx,scss,yaml,yml}'",
     "format-check": "prettier --check '**/*.{js,json,md,ts,tsx,scss,yaml,yml}'",
     "lint": "next lint --dir src --dir stories --dir .storybook --dir tests --dir scripts --dir frontend --dir lib --dir types",
     "lint-fix": "npm run lint -- --fix",
     "postinstall": "node ./scripts/postinstall.js",
-    "start:nr": "NODE_OPTIONS= '-r @newrelic/next' next start -p ${PORT:-3000}",
+    "start:nr": "NODE_OPTIONS='-r newrelic' next start -p ${PORT:-3000}",
     "start": "next start -p ${PORT:-3000}",
     "storybook": "storybook dev -p 6006",
     "storybook-build": "storybook build",
