+++ frontend/src/i18n/messages/es/index.ts
@@ -1,1 +1,6 @@
-export const messages = {};
+export const messages = {
+  Beta_alert: {
+    alert_title: "Este es un website en español.",
+    alert: "Puede leer algunas palabras en una lingua differente.",
+  },
+};
